[GemTalk/PharoGemStoneFFI#13] Enable caching of callout methods
This replaces the FFICalloutAPI>>function:library: extension method
which would recompile the method on every call with changes to
GciInterface and GciThreadSafeInterface so that the method is only
recompiled when the version of the GCI library is changed.

This also removes the need to change a core Pharo method which could
be changed, causing a conflict.